containers/payment-stream-processor/requirements.txt
azure-cosmosdb-table
kafka-python
